Implemented fixed width columns. Improved most parser test cases.
The converter now sets the column.x0 and x1 depending on the year
of the document, instead of computing it on the fly. The problem
was that some pages had fully cited text, leading to a shift of
the x0 and x1 and thus generating variable results that depended
on the content of the particular column.

Also, most test cases of a full document now assert the number of
titles of the full document.

Also, the identification of subparagraph and subline was moved to
ConverterParameters, fixing a todo.